Fixed I2C problem
Fixed the problem of I2C messages arriving during steppers
running. The TIMER2_COMPA_VECT routine took too long for TWI
events to be handeled as required. Stepper steps are now triggered
by setting an timer expired flag in the TIMER2_COMPA_VECT and
polling in the main loop - likewise for received TWI messages.

Note that this of course means that there is always the potential
of missing a single timer interrupt of timer 2 when doing longer
processing (like recalculating constants) inside the TWI handler!